Update the region in which parameters are looked for (#266)
### Feature or Bugfix
- Bugfix

### Detail
In the current version, if `internet_facing==True`, then
`cognito_urls_configs.py` and `frontend_config.py` look for parameters
in `us-east-1`.
But if `custom_domain` is configured, then the parameters
`frontend/custom_domain_name` and `userguide/custom_domain_name` are
created with the backend stack (thus in the region of the backend).
With this code changes, `cognito_urls_configs.py` and
`frontend_config.py` look for parameters in `us-east-1` only if
`internet_facing ==True` and `custom_domain==False`. If this is not the
case, it uses parameter in the backend region.