Visibility for shared tables and folder consumption (#289)
### Feature or Bugfix
- Feature

### Detail
- Added "shared" db name and access point names in the UI
- Added "shared" db name and access point names as part of the
shareObject graphQL schema
The result is that users can view relevant information to consume data
that has been shared with them directly from the shares view. They can
copy to their clipboards a command to list the objects inside a shared
folders using the AWS CLI or they can copy an SQL statement to use in
Athena and query shared tables.


![image](https://user-images.githubusercontent.com/71252798/216063148-dabb0834-a1bd-45e6-8446-916540efc871.png)


### Relates
- #283 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.