388 race condition occurs when adding folder to shared items in shares and errors out (#392)
### Feature or Bugfix
- Bugfix

### Detail
- The creation of S3 access points is asynchronous and can take more
than 5 seconds to complete. When the share managers tries attaching the
policy to the access points it fails in certain cases. This PR replaces
the waiting time of 5 seconds for a while loop that checks that the
access points has been created and if not it waits for 30s

### Relates
- #388 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.