Fix sharing update (#404)
### Feature or Bugfix
- Bugfix

### Detail
When we import a dataset stack, if the S3 location was already
registered data.all does not create a storage location. The issue is
that for datasets where data.all needs to create a storage location:
1. the first time that it creates the stack it detects that there is no
storage location and it creates the corresponding CFN resource
2. the first time that it UPDATES the stack it detects the storage
location from 1. and it deletes the CFN resource
3. the next time that it UPDATES the stack it does not detect any
storage location (it was deleted in 2.) and it creates the CFN resource
again.

To fix this behavior, in V1.5 we will use Lambda custom resource to
check the storage location and avoid CFN resources. But for previous
versions, this PR includes:
- in the method that checks the existence of an storage location, we
filter by the roleArn of the location. If the roleArn is the
`dataallPivotRole` then we assume that it was created by the dataset,
which means that `existing_storage_location = False`

I tested locally but with actual stacks being created. No additional
policies are needed

### Relates

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.