Backwards compatibility of SageMaker Studio infra + add VPC creation for SageMaker domain (#420)
### Feature or Bugfix
- Feature
- Bugfix

### Detail
- Instead of creating the SageMaker Studio domain as a nested stack we
create it as part of the environment stack. To clearly show that the
resources created for SageMaker are part of the ML Studio functionality
they `check_existing_sagemaker_studio_domain` and
`create_sagemaker_domain_resources` are class methods of
`SageMakerDomain` placed in
`backend/dataall/cdkproxy/stacks/sagemakerstudio.py`.
- As reported in #352 data.all uses the default VPC of the account,
which does not fill the requirements for SM Studio. This results in long
start times. This PR also adds the creation of a dedicated VPC that
solves the issue of slow starts.
- It is not possible to modify the networking configuration of an
existing SageMaker Studio domain. In CloudFormation it deletes and
re-creates the domain (replacement= True), and if it has Studio users it
results in failure of the CloudFormation stack. For this reason I kept
the previous implementation using the default VPC. If a customer opts to
use a dedicated networking they need to delete the default VPC. This is
an interim solution and we will look for better ways to migrate to a
dedicated SM VPC once we get more info on how customers are using
data.all ML Studio

### Relates
- #409 
- #352 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.