Fix handlers v1m5m0 + documentation (#431)
### Feature or Bugfix
- Bugfix
- Refactoring

### Detail
- Reverted LakeFormation registering location to a CloudFormation
resource instead of using a custom resource. There was a bug on the
CloudFormation LakeFormation resource and a storage location could not
be updated. The bug has been fixed and we can now update the
LakeFormation resources (which was needed to update from PivotRole to
PivotRole-cdk)
- Readded the original custom resource that creates the dataset Glue
database and grants permissions in Lake Formation. We wanted to replace
the custom resource service token because it was causing Lambda policies
to reach their maximum size; but it is not possible to update the
service token of a custom resource, one has to recreate it. The issue is
that when we directly replace the custom resource it deletes the
database (on_delete) when the CloudFormation stack cleans up the
resources. Deleting the Glue database does not delete the S3 data, but
it deletes all Lake Formation permissions, which is critical for
production use-cases. We cannot introduce this risk, so we have updated
the old custom resource Lambda to not delete anything on the on_delete
event and we will leave the deprecated custom resource until the next
minor release where we will safely remove it from the dataset stack. In
V1.6.0 we will ensure to notify that upgrading to V1.5.0 is a
requirement before upgrading to V1.6.0.
- Added documentation for ML Studio networking requirements and new
features
- Added UserGuide as pdf for V1.5.0

### Relates
- #409 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.