Fixed unresolved reference for ShareObjectPermission (#418)
### Feature or Bugfix
- Bugfix

### Detail
Working on datasets, I found that there was no `ApproversAndRequesters`
value in `ShareObjectPermission`. Whenever the code reaches this point
there will be an error.

I added the other value to avoid an exception. But this is just for
awareness.
We might need to add `ApproversAndRequesters` to `ShareObjectPermission`
I also considered that It might be dead code that we could delete.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.