Modify docker-compose yaml to read region and default region from env… (#446)
### Feature or Bugfix
- Bugfix
- Refactoring

### Detail
Added `AWS_REGION` to the environment variables of the Docker containers
for local development. Set both`AWS_DEFAULT_REGION` and `AWS_REGION` to
their values set on the terminal where `docker-compose up` is run.
If these values are not set, `eu-west-1` is used as default
Another PR with better instructions to the github pages documentation
(deploy locally) will follow.

### Relates

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.