Datasets modularization pt.5 (#442)
### Feature or Bugfix
- Refactoring

### Detail
Refactoring of the Dataset entity and related to it code.
Refactoring for Votes
Introduced DataPolicy (the same way as ServicePolicy was used used)
Extracted dataset related permissions. 
Used new `has_tenant_permission` instead of `has_tenant_perm` that
allows not to pass unused parameters

### Relates
#412 and #295

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.