fix: Upgrade aurora engine version to 11.16 (#471)
### Feature or Bugfix

- Bugfix


### Detail
Update Aurora engine version to 11.16. Fixes an issue where the Aurora
nested stack deployment in the data.all backend which goes to deployment
account would fail as AuroraPostgresEngineVersion.VER_10_18 is not
compatible with parameter group default.aurora-postgresql11
### Relates
- https://github.com/awslabs/aws-dataall/pull/466

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.