feat: limit dataset role permissions (#497)
### Feature or Bugfix
- Refactoring

### Detail
The resulting IAM policy can:
- list all buckets
- read and write objects to the dataset Bucket which is encrypted
- read S3 access points in the dataset Bucket
- putLogs in dataset Glue crawler log group
- read dataset Glue database, read and write tables in the dataset Glue
database. This is not strictly necessary as in data.all permission to
data is handled using Lake Formation. But restricting the IAM-based data
permissions we ensure that any Glue resource that is not protected using
Lake Formation is not accessible by this role
- WIP - read objects to the `/profiling/code` prefix in the environment
bucket
- WIP - read and write objects to the
`/profiling/code/results/datasetUri/` prefix in the environment bucket

IMPORTANT: I found a bug related to profiling jobs that prevented me to
test the profiling jobs. A separate
[issue](https://github.com/awslabs/aws-dataall/issues/506) has been
opened for it. For this reason the profiling permissions are a work in
progress and might require changes. e.g. additional KMS permissions.

It cannot:
- read or write to any other S3 Bucket
- use any KMS key different from the dataset KMS key
- read or write any other Glue database/tables

In addition, the Glue crawler and the profiling Job of the dataset have
been modified to always use the dataset role and not the PivotRole to
break down the "super permissions" of the pivot role and distribute
responsibilities. As a result, the dataset role can be assumed:
- by the pivotRole -> used whenever users are assuming the role from
data.all UI
- by Glue -> to run Glue profiling jobs and Glue crawler

### Relates
- #461 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: Gezim Musliaj <102723839+gmuslia@users.noreply.github.com>