dependabot - upgradefast-xml-parser, aws-amplify, react-scripts, override react-redux to non-vulnerable version (#521)
### Feature or Bugfix
- Bugfix

### Detail
- Upgraded `fast-xml-parser`
- In the process I also found that other dependency libraries included
vulnerabilities. In particular `react-redux` and `nth-check`, the parent
packages `aws-amplify`, `react-scripts` and `appbaseio/reactivesearch`
have been updated. For this last one, the latest version still uses a
vulnerable version of `react-redux` so I added a ovverride clause in the
package.json

### Relates
- Related to
https://github.com/NaturalIntelligence/fast-xml-parser/releases/tag/v4.2.4

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.