feat: limit environment roles permissions + import KMS key in imported datasets (#515)
### Feature or Bugfix
- Feature

### Detail
With this PR, the IAM policies for the environment admin group and for
environment invited groups, are generated in the same way. We remove
unnecessary duplicities and use a single method to create both.

The resulting role has been hardened and its policies restricted. The
following sheet contains the analysis of the 2 types of roles and the
measures taken to restrict the policies.


[DataallEnvRolesAnalysis.xlsx](https://github.com/awslabs/aws-dataall/files/11732639/DataallEnvRolesAnalysis.xlsx)

As a summary:
- restricted access to S3 buckets and KMS keys to the Team-owned
datasets only
- restricted access to environment s3 bucket, athenaqueries result of
the team only
- restricted trust policies to Glue and iam:PassRole to Lambda, Glue,
StepFunctions, Databrew and SageMaker
- restricted creation and access to resources of any type: all must be
tagged with tag `Team=teamname` and prefixed by the environment
specified `resource_prefix`
- removed unused Redshift policies and created new service policies
(ssm, secretsmanager, athena) to clearly differenciate policies.


To restrict access to KMS keys, we had to introduce a KMS Alias field in
the importing dataset form, allowing customers to import existing KMS
keys. The necessary database and graphql fields were already in place,
we just made it available as input of the imported dataset. It adds a
check to import the KMS resource, in case the KMS key does not exists
the stack deployment will fail. In this PR we also modify the frontend
views to show the KMS alias in imported datasets.


![image](https://github.com/awslabs/aws-dataall/assets/71252798/5ff4fdbf-43ee-47fe-a162-48da4e760575)

Userguide documentation has been updated:
1) add KMS import information in datasets
2) description of dataset IAM role - usage, policies, data access
3) description of env roles - usage, policies, data access

Finally, the CloudFormation output giving the pivotRole name has been
updated to be more specific to the environment and allow multiple
environments being deployed in the same account/region (not recommended,
but might be needed in some customers)

### Relates
- #462 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.