Resolve dataset share checks when deleting dataset (#554)
### Feature or Bugfix
- Bugfix


### Detail
- Fix query to check dataset shares before deleting a dataset

The `existingSharedItems` property of `ShareObject` was not resolving
correctly to a `True` or `False` value but rather to `None`. The impact
is:

1. When checking a dataset's shares before deleting the dataset - we
would not raise an exception for datasets that did in fact have existing
shared items.

2. Additionally, when deleting a dataset with associated shared objects
created - we would not successfully delete the associated share object
and share object items and not clean up rows in the DB appropriately.
This caused issues when viewing the Share Tab as the dataset does not
exist but is still referenced in the outdated ShareObject and throws an
error.


### Relates
- #544 


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.