Added ec2:DescribePrefix permissions to CDKSynth (#566)
### Feature or Bugfix
- BugFix

### Detail
For the release of v1.6 one permission is needed in CDK Synth CodeBuild
stage. Since this permission will be added in the next CodeBuild stage
of the deployment pipeline (UpdatePipeline and SelfMutate) the pipeline
execution will fail when upgrading to V1.6.
This PR adds the permission so if customers want to avoid manually
adding the permission they can upgrade to v1.5.X and then to V1.6.0 in a
2 steps upgrade

### Relates
Release v1.6.0

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.