v1.6.0 backwards compatibility changes (#567)
### Feature or Bugfix
- Bugfix

### Detail
- When updating the dataset database handler custom resource to use a
different PhysicalId to distinguish imported datasets CloudFormation
forces a replacement of the resource which results in the deletion of
the glue database. With this PR we add a condition to only delete
databases when the custom resource physical id is prefixed by
"CREATED-".
- Force that Glue crawlers ccreated by data.all use the dataset IAM role
as execution role. Previously a boto3 call updated this to use the
pivotRole (contrary to what it was defined in CloudFormation). With this
PR we make sure that the correct role is used in any existing crawler.

I tested in a real AWS deployment that had previously created
environments and datasets, with a crawler with the pivotRole assigned:

- [X] Crawler execution role gets updated to dataset role
- [X] Imported glue databases do not get deleted when the custom
resource updates the physicalID

### Relates
- #531 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.