SSE-S3 for imported datasets + rename imported dataset resources (#572)
### Feature or Bugfix
- Feature
- Bugfix

### Detail
#### SSE-S3 for imported datasets 
With #515 we added the possibility to introduce the KMS key that
encrypts the imported bucket when importing a dataset. This is a
required field that cannot be empty. In this PR we allow customers to
leave the field empty for the case in which no KMS key is used. We
assume that if no key is defined, then the bucket is encrypted using
SSE-S3 encryption.

- changes in frontend views
- changes in definition of RDS column values
- added line in migration script to backfill `undefined` keys with the
`undefined` value for previously imported datasets.

#### Rename imported dataset resources
In #535 the permissions given to the pivot role were restricted to more
specific resources. Most policies where updated to access
`resource-prefix` AWS resources only. For imported datasets some of the
data.all created resources in the dataset stack (Glue crawler ....) do
not follow this naming pattern. In this PR we update the name of those
resources to include the prefix. This way the pivot role will be able to
execute actions on them.

- Added naming convention pattern
- Enforce pattern in dataset import
- Added migration script to update the names of existing imported
datasets

⚠️ After the migration script is executed, dataset stacks need to be
updated for the AWS resources to be updated ⚠️

### Testing - in-progress
I tested locally, but now I am also testing in a real deployment:
- [ ] CICD pipeline completes successfully
- [ ] KMS RDS info updated as expected
- [ ] KMS RDS info of new created and imported datasets as expected
- [ ] CloudFormation stacks of updated stacks of imported and created
datasets successful (resources updated)
- [ ] CloudFormation stacks of new stacks of imported and created
datasets successful (resources prefixed)
- [ ] Crawlers can be run for all dataset cases

### Relates
- #535
- #515

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.