fix: Adding cookiecutter config and environment variable for datapipelines stacks (#582)
Workaround for cookiecutter file writes to root filesystem, which occurs
during creation of datapipelines.

### Feature or Bugfix
- Bugfix

### Detail
Cookiecutter is used by DDK to generate a sample project. However it tries to write to the root file system which conflicts with our "read-only" settings on root. This PR modifies the settings of cookiecutter to write to `/dataall` instead

### Relates
- #523 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Co-authored-by: dlpzx <dlpzx@amazon.com>