Migrate to a new permission checker (#569)
There is a new permission checker API in the modularization branch. It
allows us get away from passing unused variables into the methods.
The previous permission checker used a certain signature that expected
special parameters passed in a certain order. Though, most of the time
those passed parameters were not used in the method, but only were
needed for the permission checker decorator.
The other issue with the previous permission checker, that it had forced
us to pass a request data as a dictionary e.g. in the resolver we had
had some parameters and had to put them into dictionary and unpack them
back in the method. It was not memory inefficient
All in all, a new permission checker doesn't require to follow previous
agreements which will us to write shorter and more concise API, we can
utilize type checking and avoid unnecessary memory allocation during the
object creation.


The PR is waiting other PR to merged.


By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.

---------

Signed-off-by: dependabot[bot] <support@github.com>
Co-authored-by: dlpzx <71252798+dlpzx@users.noreply.github.com>
Co-authored-by: dependabot[bot] <49699333+dependabot[bot]@users.noreply.github.com>
Co-authored-by: Gezim Musliaj <102723839+gmuslia@users.noreply.github.com>
Co-authored-by: kukushking <kukushkin.anton@gmail.com>
Co-authored-by: Rick Bernotas <97474536+rbernotas@users.noreply.github.com>
Co-authored-by: David Mutune Kimengu <57294718+kimengu-david@users.noreply.github.com>
Co-authored-by: dbalintx <132444646+dbalintx@users.noreply.github.com>