modularization: pivot role policies (#616)
### Feature or Bugfix
- Feature and refactoring

### Detail

Implement a class to add IAM Policy statements to the pivotRole policies
from the modules. The additional pivot role policies needed for each
module can be defined inside the module code as in the following
picture:
<img width="1264" alt="image"
src="https://github.com/awslabs/aws-dataall/assets/71252798/d82d2411-7639-494b-9783-f3e17a0d633d">

Policies that are needed independently from the modules are defined in
`dataall/core/environment/cdk/pivot_role_core_policies`.

I moved all permissions, including those for datasets and data_sharing
inside their modules.

In addition:
- Splitted MLStudio and Notebooks pivot role permissions
- Renamed policies added to the environment team roles and moved them to
`core/environment/cdk`

### Questions
- Some modules share permissions so there is some code-duplication
(around 10 lines). In my opinion it is minimal, and it is not worth
mixing module permissions, but I wanted to bring it up.
- With this PR we can only update based on modules the
`dataallPivotRole-cdk` that is created as part of the environment stack
when auto-create is set to true. Should we implement something for
manually created pivot roles? These are roles that are manually deployed
in the environment account and cannot be handled by data.all.


### Relates
- #610 

By submitting this pull request, I confirm that my contribution is made
under the terms of the Apache 2.0 license.